API change: entity position are integral at cell centers
This change breaks existing level sets, so all existing
level sets were updated to the new standard. A few level
bugs may result from this.

The coordinate system used in the main code is unchanged;
the guile interface adds or subtracts 0.5 to the X and Y
coordinates as appropriate.

As a side effect of this, all animated objects can be placed
at real coordinates.

The (add-flag...) method has been adjusted so its 'visible'
argument must now be a boolean.

There were a few minor fixes in the process.