Update maps for removed texture and limit object creation
An image was removed in an earlier commit, but the level maps
held references to it, which lead to an unneeded error message.

Also, some levels created objects inside the function that
`trigger` calls, leading to duplicated objects and lag as
trigger would run several hundred times. The new
`trigger-once` defined in `boot.scm` only calls the function
the first time.