Perform string translations at the last minute
Level names are stored in the original (English)
and translation is deferred until the text is
used on screen or in game. Previously the save
files would record translated level names, and
levels' names would only adjust to language changes
on playing.

This change is backwards compatible with existing
highscores and player data.

To permit this change, the translation function `_`
in the level scripts does not alter the string passed
in; the (set-author ...) and (sign ...) translate
their string arguments if necessary.