Add options for shadow detail, sandbox modes, and time scaling
Shadow detail controls texture size for shadows. The difficulty
options for sandbox mode are not only available if the associated
flag is active, as sandbox mode is better for testing than gameplay.

Time scaling speeds up or slows down the apparent system time, so
that physics and other game code can be run at half or at twice the
standard frequency, depending on scaling constant.