Clarify license in README
All the source code is definitely provided under GPL v2 or later,
as per the comments at the top. Everything else* is under the "GPL",
with no version number stated, which -- as explicitly stated in
each version so far of the GPL -- means that it is provided under
any version of the GPL, and thus can be distributed under GPLv2-or-later.

* The one exception is the music in share/music, some of which
falls under the Ethymos Free Music License. If you wish to distribute
a pure GPLv2 or later version of trackballs, just delete the non-GPL
files. This won't break anything, because trackballs automatically plays
whatever ogg files are in the share/music/ folder. Note that by default,
trackballs installs the EFML license in the README alongside the music
itself.